refactor(strategy): 이동전략을 strategy 패키지로 분리
