+++ src/main/java/racinggame/domain/car/RacingCars.java
@@ -1,5 +1,7 @@
 package racinggame.domain.car;
 
+import racinggame.domain.strategy.MovableStrategy;
+
 import java.util.Arrays;
 import java.util.Comparator;
 import java.util.List;
