+++ src/main/java/racinggame/domain/strategy/RandomMovableStrategy.java
@@ -1,4 +1,4 @@
-package racinggame.domain.car;
+package racinggame.domain.strategy;
 
 import java.util.Random;
